Renamed react component
